+++ src/Serilog.Seq/SerilogSeqSinkHelper.cs
@@ -5,7 +5,9 @@
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
+using Serilog.Formatting;
 using Serilog.Sinks.PeriodicBatching;
+using Serilog.Sinks.Seq;
 
 namespace Phoenix.Functionality.Logging.Extensions.Serilog.Seq;
 
@@ -13,35 +15,37 @@ internal class SerilogSeqSinkHelper
 {
     internal delegate bool EvaluationFunction(LogEvent logEvent);
 
-    /// <summary>
-    /// Creates an <see cref="ILogEventSink"/> capable to write to a seq server.
-    /// </summary>
-    /// <param name="seqHost"> The seq host (e.g. https://localhost or https://localhost:5341). </param>
-    /// <param name="seqPort"> An optional port of the seq server. The port can also be specified in <paramref name="seqHost"/>. </param>
-    /// <param name="applicationTitle"> The title of the application that will log to seq. </param>
-    /// <param name="configurationApiKey"> Api key used for configuration. </param>
-    /// <param name="retryOnError"> Should registering the application with the seq server automatically be retried, if it initially failed. </param>
-    /// <param name="retryCount"> The amount of attempts made to register the application after the initial one failed. If this is null, then endless attempts will be made. Is only used if <paramref name="retryOnError"/> is true. </param>
-    /// <param name="controlLevelSwitch"> If provided, the switch will be updated based on the Seq server's level setting for the corresponding API key. Passing the same key to MinimumLevel.ControlledBy() will make the whole pipeline dynamically controlled. Do not specify restrictedToMinimumLevel with this setting. </param>
-    /// <param name="batchPostingLimit"> The maximum number of events to post in a single batch. </param>
-    /// <param name="period"> The time to wait between checking for event batches. </param>
-    /// <param name="eventBodyLimitBytes"> The maximum size, in bytes, that the JSON representation of an event may take before it is dropped rather than being sent to the Seq server. Specify null for no limit. The default is 265 KB. </param>
-    /// <param name="messageHandler"> Used to construct the HttpClient that will send the log messages to Seq. </param>
-    /// <param name="queueSizeLimit"> The maximum number of events that will be held in-memory while waiting to ship them to Seq. Beyond this limit, events will be dropped. The default is 5000. Has no effect on durable log shipping. </param>
-    /// <param name="selfLogger"> Optional <see cref="SelfLogger"/> used for internal logging. Default is <see cref="SelfLogger.DefaultSelfLogger"/>. </param>
-    /// <returns>
-    /// <para> A <see cref="ValueTuple"/> </para>
-    /// <para> <see cref="Nullable{ILogEventSink}"/> Sink: The sink capable to emit <see cref="LogEvent"/>s to a seq server. </para>
-    /// <para> <see cref="Nullable{EvaluationFunction}"/> EvaluationFunction: The evaluation function that should be used with <see cref="global::Serilog.Configuration.LoggerSinkConfiguration.Conditional"/>. </para>
-    /// </returns>
-    internal static (ILogEventSink? Sink, EvaluationFunction? EvaluationFunction) CreateSink
+	/// <summary>
+	/// Creates an <see cref="ILogEventSink"/> capable to write to a seq server.
+	/// </summary>
+	/// <param name="seqHost"> The seq host (e.g. https://localhost or https://localhost:5341). </param>
+	/// <param name="seqPort"> An optional port of the seq server. The port can also be specified in <paramref name="seqHost"/>. </param>
+	/// <param name="applicationTitle"> The title of the application that will log to seq. </param>
+	/// <param name="configurationApiKey"> Api key used for configuration. </param>
+	/// <param name="retryOnError"> Should registering the application with the seq server automatically be retried, if it initially failed. </param>
+	/// <param name="retryCount"> The amount of attempts made to register the application after the initial one failed. If this is null, then endless attempts will be made. Is only used if <paramref name="retryOnError"/> is true. </param>
+	/// <param name="payloadFormatter"> An <see cref="ITextFormatter"/> that will be used to format (newline-delimited CLEF/JSON) payloads. </param>
+	/// <param name="controlLevelSwitch"> If provided, the switch will be updated based on the Seq server's level setting for the corresponding API key. Passing the same key to MinimumLevel.ControlledBy() will make the whole pipeline dynamically controlled. Do not specify restrictedToMinimumLevel with this setting. </param>
+	/// <param name="batchPostingLimit"> The maximum number of events to post in a single batch. </param>
+	/// <param name="period"> The time to wait between checking for event batches. </param>
+	/// <param name="eventBodyLimitBytes"> The maximum size, in bytes, that the JSON representation of an event may take before it is dropped rather than being sent to the Seq server. Specify null for no limit. The default is 265 KB. </param>
+	/// <param name="messageHandler"> Used to construct the HttpClient that will send the log messages to Seq. </param>
+	/// <param name="queueSizeLimit"> The maximum number of events that will be held in-memory while waiting to ship them to Seq. Beyond this limit, events will be dropped. The default is 5000. Has no effect on durable log shipping. </param>
+	/// <param name="selfLogger"> Optional <see cref="SelfLogger"/> used for internal logging. Default is <see cref="SelfLogger.DefaultSelfLogger"/>. </param>
+	/// <returns>
+	/// <para> A <see cref="ValueTuple"/> </para>
+	/// <para> <see cref="Nullable{ILogEventSink}"/> Sink: The sink capable to emit <see cref="LogEvent"/>s to a seq server. </para>
+	/// <para> <see cref="Nullable{EvaluationFunction}"/> EvaluationFunction: The evaluation function that should be used with <see cref="global::Serilog.Configuration.LoggerSinkConfiguration.Conditional"/>. </para>
+	/// </returns>
+	internal static (ILogEventSink? Sink, EvaluationFunction? EvaluationFunction) CreateSink
     (
         string seqHost,
         ushort? seqPort,
         string applicationTitle,
-        string? configurationApiKey = null,
+		string? configurationApiKey = null,
         bool retryOnError = true,
         byte? retryCount = null,
+		ITextFormatter? payloadFormatter = null,
         LoggingLevelSwitch? controlLevelSwitch = null,
         int batchPostingLimit = 1000,
         TimeSpan? period = null,
@@ -50,12 +54,13 @@ internal class SerilogSeqSinkHelper
         int queueSizeLimit = 5000,
         SelfLogger? selfLogger = null
     ) =>
-        SerilogSeqSinkHelper.CreateSink
+        CreateSink
         (
             new SeqServer(seqHost, seqPort, configurationApiKey),
             applicationTitle,
             retryOnError,
             retryCount,
+			payloadFormatter,
             controlLevelSwitch,
             batchPostingLimit,
             period,
@@ -65,31 +70,33 @@ internal class SerilogSeqSinkHelper
             selfLogger
         );
 
-    /// <summary>
-    /// Creates an <see cref="ILogEventSink"/> capable to write to a seq server.
-    /// </summary>
-    /// <param name="seqServer"> The <see cref="SeqServer"/> used to register the application. </param>
-    /// <param name="applicationTitle"> The title of the application that will log to seq. </param>
-    /// <param name="retryOnError"> Should registering the application with the seq server automatically be retried, if it initially failed. </param>
-    /// <param name="retryCount"> The amount of attempts made to register the application after the initial one failed. If this is null, then endless attempts will be made. Is only used if <paramref name="retryOnError"/> is true. </param>
-    /// <param name="controlLevelSwitch"> If provided, the switch will be updated based on the Seq server's level setting for the corresponding API key. Passing the same key to MinimumLevel.ControlledBy() will make the whole pipeline dynamically controlled. Do not specify restrictedToMinimumLevel with this setting. </param>
-    /// <param name="batchPostingLimit"> The maximum number of events to post in a single batch. </param>
-    /// <param name="period"> The time to wait between checking for event batches. </param>
-    /// <param name="eventBodyLimitBytes"> The maximum size, in bytes, that the JSON representation of an event may take before it is dropped rather than being sent to the Seq server. Specify null for no limit. The default is 265 KB. </param>
-    /// <param name="messageHandler"> Used to construct the HttpClient that will send the log messages to Seq. </param>
-    /// <param name="queueSizeLimit"> The maximum number of events that will be held in-memory while waiting to ship them to Seq. Beyond this limit, events will be dropped. The default is 5000. Has no effect on durable log shipping. </param>
-    /// <param name="selfLogger"> Optional <see cref="SelfLogger"/> used for internal logging. Default is <see cref="SelfLogger.DefaultSelfLogger"/>. </param>
-    /// <returns>
-    /// <para> A <see cref="ValueTuple"/> </para>
-    /// <para> <see cref="Nullable{ILogEventSink}"/> Sink: The sink capable to emit <see cref="LogEvent"/>s to a seq server. </para>
-    /// <para> <see cref="Nullable{EvaluationFunction}"/> EvaluationFunction: The evaluation function that should be used with <see cref="global::Serilog.Configuration.LoggerSinkConfiguration.Conditional"/>. </para>
-    /// </returns>
-    internal static (ILogEventSink? Sink, EvaluationFunction? EvaluationFunction) CreateSink
+	/// <summary>
+	/// Creates an <see cref="ILogEventSink"/> capable to write to a seq server.
+	/// </summary>
+	/// <param name="seqServer"> The <see cref="SeqServer"/> used to register the application. </param>
+	/// <param name="applicationTitle"> The title of the application that will log to seq. </param>
+	/// <param name="retryOnError"> Should registering the application with the seq server automatically be retried, if it initially failed. </param>
+	/// <param name="retryCount"> The amount of attempts made to register the application after the initial one failed. If this is null, then endless attempts will be made. Is only used if <paramref name="retryOnError"/> is true. </param>
+	/// <param name="payloadFormatter"> An <see cref="ITextFormatter"/> that will be used to format (newline-delimited CLEF/JSON) payloads. </param>
+	/// <param name="controlLevelSwitch"> If provided, the switch will be updated based on the Seq server's level setting for the corresponding API key. Passing the same key to MinimumLevel.ControlledBy() will make the whole pipeline dynamically controlled. Do not specify restrictedToMinimumLevel with this setting. </param>
+	/// <param name="batchPostingLimit"> The maximum number of events to post in a single batch. </param>
+	/// <param name="period"> The time to wait between checking for event batches. </param>
+	/// <param name="eventBodyLimitBytes"> The maximum size, in bytes, that the JSON representation of an event may take before it is dropped rather than being sent to the Seq server. Specify null for no limit. The default is 265 KB. </param>
+	/// <param name="messageHandler"> Used to construct the HttpClient that will send the log messages to Seq. </param>
+	/// <param name="queueSizeLimit"> The maximum number of events that will be held in-memory while waiting to ship them to Seq. Beyond this limit, events will be dropped. The default is 5000. Has no effect on durable log shipping. </param>
+	/// <param name="selfLogger"> Optional <see cref="SelfLogger"/> used for internal logging. Default is <see cref="SelfLogger.DefaultSelfLogger"/>. </param>
+	/// <returns>
+	/// <para> A <see cref="ValueTuple"/> </para>
+	/// <para> <see cref="Nullable{ILogEventSink}"/> Sink: The sink capable to emit <see cref="LogEvent"/>s to a seq server. </para>
+	/// <para> <see cref="Nullable{EvaluationFunction}"/> EvaluationFunction: The evaluation function that should be used with <see cref="global::Serilog.Configuration.LoggerSinkConfiguration.Conditional"/>. </para>
+	/// </returns>
+	internal static (ILogEventSink? Sink, EvaluationFunction? EvaluationFunction) CreateSink
     (
         SeqServer seqServer,
         string applicationTitle,
         bool retryOnError = true,
         byte? retryCount = null,
+		ITextFormatter? payloadFormatter = null,
         LoggingLevelSwitch? controlLevelSwitch = null,
         int batchPostingLimit = 1000,
         TimeSpan? period = null,
@@ -97,7 +104,7 @@ internal class SerilogSeqSinkHelper
         HttpMessageHandler? messageHandler = null,
         int queueSizeLimit = 5000,
         SelfLogger? selfLogger = null
-    )
+	)
     {
         selfLogger ??= SelfLogger.DefaultSelfLogger;
         
@@ -127,7 +134,7 @@ internal class SerilogSeqSinkHelper
         }
 
         // Get the seq requirements.
-        var couldGetSeqRequirements = SerilogSeqSinkHelper.TryGetSeqRequirements(out var seqSink, out var evaluationFunction, selfLogger, seqServer.ConnectionData.Url, apiKey, controlLevelSwitch, eventBodyLimitBytes, messageHandler);
+        var couldGetSeqRequirements = TryGetSeqRequirements(out var seqSink, out var evaluationFunction, selfLogger, seqServer.ConnectionData.Url, apiKey, payloadFormatter, controlLevelSwitch, eventBodyLimitBytes, messageHandler);
         if (!couldGetSeqRequirements)
         {
             selfLogger.Log($"Could not create the required seq objects.");
@@ -135,7 +142,7 @@ internal class SerilogSeqSinkHelper
         }
 
         // Build the periodic batching sink that wraps the seq sink.
-        var periodicBatchingSink = SerilogSeqSinkHelper.GetPeriodicBatchingSink(seqSink!, batchPostingLimit, period, queueSizeLimit);
+        var periodicBatchingSink = GetPeriodicBatchingSink(seqSink!, batchPostingLimit, period, queueSizeLimit);
 
         if (couldRegisterApplication)
         {
@@ -148,21 +155,24 @@ internal class SerilogSeqSinkHelper
     }
 
 #if NETSTANDARD2_0 || NETSTANDARD1_6 || NETSTANDARD1_5 || NETSTANDARD1_4 || NETSTANDARD1_3 || NETSTANDARD1_2 || NETSTANDARD1_1 || NETSTANDARD1_0
-    internal static bool TryGetSeqRequirements(out IBatchedLogEventSink? seqSink, out EvaluationFunction? evaluationFunction, SelfLogger selfLogger, string serverUrl, string? apiKey = null, LoggingLevelSwitch? controlLevelSwitch = null, long? eventBodyLimitBytes = 256 * 1024, System.Net.Http.HttpMessageHandler? messageHandler = null)
+    internal static bool TryGetSeqRequirements(out IBatchedLogEventSink? seqSink, out EvaluationFunction? evaluationFunction, SelfLogger selfLogger, string serverUrl, string? apiKey = null, ITextFormatter? payloadFormatter = null, LoggingLevelSwitch? controlLevelSwitch = null, long? eventBodyLimitBytes = 256 * 1024, System.Net.Http.HttpMessageHandler? messageHandler = null)
 #else
-		internal static bool TryGetSeqRequirements([System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out IBatchedLogEventSink? seqSink, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out EvaluationFunction? evaluationFunction, SelfLogger selfLogger, string serverUrl, string? apiKey = null, LoggingLevelSwitch? controlLevelSwitch = null, long? eventBodyLimitBytes = 256 * 1024, System.Net.Http.HttpMessageHandler? messageHandler = null)
+	internal static bool TryGetSeqRequirements([System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out IBatchedLogEventSink? seqSink, [System.Diagnostics.CodeAnalysis.NotNullWhen(true)] out EvaluationFunction? evaluationFunction, SelfLogger selfLogger, string serverUrl, string? apiKey = null, ITextFormatter? payloadFormatter = null, LoggingLevelSwitch? controlLevelSwitch = null, long? eventBodyLimitBytes = 256 * 1024, System.Net.Http.HttpMessageHandler? messageHandler = null)
 #endif
     {
         evaluationFunction = null;
         seqSink = null;
 
-        var controlledLevelSwitch = SerilogSeqSinkHelper.GetControlledLevelSwitch(selfLogger, controlLevelSwitch);
+        var controlledLevelSwitch = GetControlledLevelSwitch(selfLogger, controlLevelSwitch);
         if (controlledLevelSwitch is null) return false;
-        evaluationFunction = SerilogSeqSinkHelper.GetEvaluationFunction(selfLogger, controlledLevelSwitch);
+        
+		evaluationFunction = GetEvaluationFunction(selfLogger, controlledLevelSwitch);
         if (evaluationFunction is null) return false;
-        var ingestionApi = SerilogSeqSinkHelper.GetIngestionApiClient(selfLogger, serverUrl, apiKey, messageHandler);
+        
+		var ingestionApi = GetIngestionApiClient(selfLogger, serverUrl, apiKey, messageHandler);
         if (ingestionApi is null) return false;
-        seqSink = SerilogSeqSinkHelper.GetSeqSink(selfLogger, ingestionApi, controlledLevelSwitch, eventBodyLimitBytes);
+        
+		seqSink = GetSeqSink(selfLogger, ingestionApi, payloadFormatter ?? new SeqCompactJsonFormatter(), controlledLevelSwitch, eventBodyLimitBytes);
         if (seqSink is null) return false;
 
         return true;
@@ -179,7 +189,7 @@ internal class SerilogSeqSinkHelper
             var seqAssembly = typeof(SeqLoggerConfigurationExtensions).Assembly;
 
             // Get and build the ControlledLevelSwitch class.
-            var parameterTypes = new Type[]
+            var parameterTypes = new[]
             {
                 typeof(LoggingLevelSwitch)
             };
@@ -190,7 +200,7 @@ internal class SerilogSeqSinkHelper
                 selfLogger.Log($"Could not get the constructor of the '{className}' class needed to build the sink.");
                 return null;
             }
-            var instance = constructor.Invoke(new object?[] { controlLevelSwitch });
+            var instance = constructor.Invoke([controlLevelSwitch]);
             return instance;
         }
         catch (Exception ex)
@@ -208,7 +218,7 @@ internal class SerilogSeqSinkHelper
         try
         {
             // Get and build the IsIncluded function.
-            var parameterTypes = new Type[]
+            var parameterTypes = new[]
             {
                 typeof(LogEvent)
             };
@@ -222,7 +232,7 @@ internal class SerilogSeqSinkHelper
 
             bool EvaluationFunction(LogEvent logEvent)
             {
-                var result = function.Invoke(controlledLevelSwitch, new object[] { logEvent });
+                var result = function.Invoke(controlledLevelSwitch, [logEvent]);
                 return (bool?) result ?? false;
             }
 
@@ -246,7 +256,7 @@ internal class SerilogSeqSinkHelper
             var seqAssembly = typeof(SeqLoggerConfigurationExtensions).Assembly;
 
             // Get and build the SeqSink class.
-            var parameterTypes = new Type[]
+            var parameterTypes = new[]
             {
                 serverUrl.GetType(),
                 apiKey?.GetType() ?? typeof(string),
@@ -260,7 +270,7 @@ internal class SerilogSeqSinkHelper
                 selfLogger.Log($"Could not get the constructor of the '{className}' class needed to build the sink.");
                 return null;
             }
-            var instance = constructor.Invoke(new object?[] { serverUrl, apiKey, messageHandler });
+            var instance = constructor.Invoke([serverUrl, apiKey, messageHandler]);
             return instance;
         }
         catch (Exception ex)
@@ -270,7 +280,7 @@ internal class SerilogSeqSinkHelper
         }
     }
 
-    private static IBatchedLogEventSink? GetSeqSink(SelfLogger selfLogger, object ingestionApi, object controlledLevelSwitch, long? eventBodyLimitBytes = 256 * 1024)
+    private static IBatchedLogEventSink? GetSeqSink(SelfLogger selfLogger, object ingestionApi, object payloadFormatter, object controlledLevelSwitch, long? eventBodyLimitBytes = 256 * 1024)
     {
         //! This has to be done via reflection, as the 'BatchedSeqSink' class is internal.
         var className = "Serilog.Sinks.Seq.Batched.BatchedSeqSink";
@@ -281,10 +291,11 @@ internal class SerilogSeqSinkHelper
             var seqAssembly = typeof(SeqLoggerConfigurationExtensions).Assembly;
 
             // Get and build the SeqSink class.
-            var parameterTypes = new Type[]
+            var parameterTypes = new[]
             {
                 ingestionApi.GetType(),
-                eventBodyLimitBytes?.GetType() ?? typeof(long?),
+				payloadFormatter.GetType(),
+				eventBodyLimitBytes?.GetType() ?? typeof(long?),
                 controlledLevelSwitch.GetType(),
             };
 
@@ -295,7 +306,7 @@ internal class SerilogSeqSinkHelper
                 selfLogger.Log($"Could not get the constructor of the '{className}' class needed to build the sink.");
                 return null;
             }
-            var seqSink = constructor.Invoke(new object?[] { ingestionApi, eventBodyLimitBytes, controlledLevelSwitch });
+            var seqSink = constructor.Invoke([ingestionApi, payloadFormatter, eventBodyLimitBytes, controlledLevelSwitch]);
             return (IBatchedLogEventSink)seqSink;
         }
         catch (Exception ex)